server/requirements.txt
fastapi>=0.100.0 # Use a recent version
uvicorn[standard]>=0.20.0 # Includes standard dependencies like websockets
supabase>=1.0.0 # Use a recent version of the supabase client
pydantic>=2.0.0 # Required by FastAPI, ensure v2+
python-dotenv>=1.0.0 # For loading .env files

# Testing
pytest>=7.0.0
httpx>=0.24.0 # For making async requests in tests

# Agno and related dependencies
agno>=1.1.13
firecrawl>=0.1.0
anthropic>=0.8.0
ollama
crawl4ai
wikipedia
google-api-python-client
youtube-transcript-api